First Implementation of Game Engine
The game engine class creates a window and rendering context.
It manages the state of the program, whether it is in menus, in game, etc.
This is done using the GameState interface in gamestate.h. Derived
classes of GameState must implement their own HandleEvents, Render, and
Update Functions. Then they use the GameEngine class
push/pop/changeState functions to transfer states.