Created GLShaderProgram class.
GLShaderProgram is a class that encapsulates a vertex+fragment shader
program. It is capable of reading either two strings with the complete
source code or two file names, from which it will extract the source. It
then compiles, and links the code together. You can then use the shader
in the rendering pipeline via the BindProgram or UnbindProgram calls.