Implemented the logger.
This logger can be used with little or no performance hit. Just include
the header file, and use LOG(level) << "Error Msg"; in your code.
It will write to a log file or stderr depending on the reporting error
level.

There are 5 levels, and you can configure the logger to log only the
messages of a certain severity or worse. If you set the logger's
reporting level to LOG_WARNING, it will only append to the log when it
encounters a LOG request of level WARNING or worse - ERROR, FATAL. These
severity levels are defined when you call it.