First implementation of FontRenderer.
The font renderer is capable of reading a ttf font using the SDL_TTF
library.

One then calls RenderToGLTexture with the string and the color
in 0-255 RGB format, and it populates an owned texture with the text.
Then you call RenderTexture, and it binds the texture buffer containing
the text to the OpenGL state machine, so any object drawn after that
will be mapped with that texture.

It does not handle any quads, so this may be mapped in a UI Element
class.