First implementation of Quad Primitive.
The quad primitive wraps the calls for initializing a vertex array
object and vertex buffer object within a simple interface. Just call
Init() and pass the corresponding position, color, and textcoordinate
attributes from your shader, and it will associate that particular
shader with that quad. The color and textcoordinate attributes don't
need to exist, either, and the quad will associate itself just fine.

To render the quad, just activate the associated shader and attributes,
then call the Render function in the render loop.